+++ src/security/log-sanitizer.test.ts
