+++ src/utils/docs-retrieval/content-fetcher.ts
@@ -20,7 +20,7 @@ import {
   isGitHubRawUrl,
   isAllowedGitHubRawUrl,
   isLocalhostUrl,
-} from '../security';
+} from '../../security';
 import { convertGitHubRawToProxyUrl, isDataProxyUrl } from '../data-proxy';
 import { isDevModeEnabledGlobal } from '../dev-mode';
