+++ src/utils/link-handler.hook.ts
@@ -8,7 +8,13 @@ import {
   enrichWithStepContext,
 } from '../lib/analytics';
 import { getJourneyProgress } from './docs-retrieval/learning-journey-helpers';
-import { parseUrlSafely, isAllowedContentUrl, isAllowedGitHubRawUrl, isAnyGitHubUrl, isLocalhostUrl } from './security';
+import {
+  parseUrlSafely,
+  isAllowedContentUrl,
+  isAllowedGitHubRawUrl,
+  isAnyGitHubUrl,
+  isLocalhostUrl,
+} from '../security';
 import { ALLOWED_GITHUB_REPOS } from '../constants';
 import { isDevModeEnabledGlobal } from './dev-mode';
